Improve PLL calculation for BM1366/BM1368/BM1370 (#1051)
* Accept and store float frequency values

* Unify PLL resolver for 1366/1368/1370

Pick the closest frequency to the requested target. If there are multiple PLL settings, pick the one with first the lowest VDO frequency, and finally the lowest posdividers.

Merged frequency ramp start and loop. This solves overshooting when requesting a target frequency within the current step size. It also eliminates the duplicate setting of the target frequency is that's on a step boundary.

Cleaned up several unused defines.

* Restore postdiv2 < postdiv1 condition for BM1368 and BM1370

* Whitespace

* Clean up intermediate functions

Simplify frequency_transition further, finally fix overshoot

* Code cleanup

* Add debug logging for BM1397

* Fix log

* Flip postdiv condition for readability

* Add test

* Fix BM1370 fb_range

* EOF line in test_pll.c

+++ components/asic/include/bm1397.h
@@ -23,7 +23,7 @@ typedef struct __attribute__((__packed__))
     uint8_t midstate3[32];
 } job_packet;
 
-uint8_t BM1397_init(uint64_t frequency, uint16_t asic_count, uint16_t difficulty);
+uint8_t BM1397_init(float frequency, uint16_t asic_count, uint16_t difficulty);
 void BM1397_send_work(void * GLOBAL_STATE, bm_job * next_bm_job);
 void BM1397_set_version_mask(uint32_t version_mask);
 int BM1397_set_max_baud(void);
